backend/requirements.txt
# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.3
Werkzeug==2.3.7
gunicorn==21.2.0

# AI & Multi-Agent Framework
pydantic==2.7.4
typing-extensions==4.8.0
langchain==0.3.27
langchain-core==0.3.72

google-generativeai==0.8.0
langgraph==0.5.4
langgraph-swarm==0.0.13
langgraph-checkpoint==2.1.1

langgraph-supervisor==0.0.28
langchain-google-genai==2.0.1

# Database
pymongo==4.6.0
redis==5.0.1
dnspython==2.4.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Logging & Security
structlog==23.2.0
colorama==0.4.6
cryptography==41.0.7
bcrypt==4.1.2
